mise à jour de l'autoloader et de la sécurité

+++ App/Controller/Backend/DisplayController.php
@@ -8,14 +8,23 @@ class DisplayController extends Controller {
 	
 	static function displayChapter($id) {
 
-		/*verifier que l'id existe bien*/
-
 		$db = Manager::getDatabase();
 		$manager = new ChapterManager();
-		$chapter = $manager->getChapter($db, $id);
-		$content = html_entity_decode($chapter->content());
 
-		echo "<h3>" . $chapter->title() . "</h3>" . $content;
+		if ($manager->idExist($db, $id)) {
+			
+			$chapter = $manager->getChapter($db, $id);
+			$content = html_entity_decode($chapter->content());
+
+			echo "<h3>" . $chapter->title() . "</h3>" . $content;
+
+		} else {
+
+			$session->setFlash('danger', "Oups... Ce chapitre n'existe pas encore...");
+
+			header('Location: ./backindex.php?controller=edit');
+		}
+		
 
 	}
 }
